+++ src/GitHub_Updater/Basic_Auth_Loader.php
@@ -134,8 +134,8 @@ class Basic_Auth_Loader {
 		);
 
 		$repos = array_merge(
-			Singleton::get_instance( 'Plugin' )->get_plugin_configs(),
-			Singleton::get_instance( 'Theme' )->get_theme_configs()
+			Singleton::get_instance( 'Plugin', $this )->get_plugin_configs(),
+			Singleton::get_instance( 'Theme', $this )->get_theme_configs()
 		);
 
 		$slug = isset( $_REQUEST['slug'] ) ? $_REQUEST['slug'] : false;