+++ src/GitHub_Updater/Messages.php
@@ -64,7 +64,7 @@ class Messages extends Base {
 					add_action( 'network_admin_notices', array( &$this, 'show_wp_error' ) );
 					break;
 				case 'gitlab':
-					Singleton::get_instance( 'API\GitLab_API', new \stdClass() )->gitlab_error_notices();
+					Singleton::get_instance( 'API\GitLab_API', $this, new \stdClass() )->gitlab_error_notices();
 				case 'git':
 				default:
 					add_action( 'admin_notices', array( &$this, 'show_403_error_message' ) );
@@ -83,7 +83,7 @@ class Messages extends Base {
 	 */
 	public function show_403_error_message() {
 		$_403       = false;
-		$error_code = Singleton::get_instance( 'API_PseudoTrait' )->get_error_codes();
+		$error_code = Singleton::get_instance( 'API_PseudoTrait', $this )->get_error_codes();
 		foreach ( (array) $error_code as $repo ) {
 			if ( ! $_403 && 403 === $repo['code'] && 'github' === $repo['git'] ) {
 				$_403 = true;
@@ -122,7 +122,7 @@ class Messages extends Base {
 	 */
 	public function show_401_error_message() {
 		$_401       = false;
-		$error_code = Singleton::get_instance( 'API_PseudoTrait' )->get_error_codes();
+		$error_code = Singleton::get_instance( 'API_PseudoTrait', $this )->get_error_codes();
 		foreach ( (array) $error_code as $repo ) {
 			if ( ! $_401 && 401 === $repo['code'] ) {
 				$_401 = true;