+++ src/GitHub_Updater/Rest_Update.php
@@ -62,7 +62,7 @@ class Rest_Update extends Base {
 		$plugin           = null;
 		$is_plugin_active = false;
 
-		foreach ( (array) Singleton::get_instance( 'Plugin' )->get_plugin_configs() as $config_entry ) {
+		foreach ( (array) Singleton::get_instance( 'Plugin', $this )->get_plugin_configs() as $config_entry ) {
 			if ( $config_entry->repo === $plugin_slug ) {
 				$plugin = $config_entry;
 				break;
@@ -113,7 +113,7 @@ class Rest_Update extends Base {
 	public function update_theme( $theme_slug, $tag = 'master' ) {
 		$theme = null;
 
-		foreach ( (array) Singleton::get_instance( 'Theme' )->get_theme_configs() as $config_entry ) {
+		foreach ( (array) Singleton::get_instance( 'Theme', $this )->get_theme_configs() as $config_entry ) {
 			if ( $config_entry->repo === $theme_slug ) {
 				$theme = $config_entry;
 				break;
@@ -356,7 +356,7 @@ class Rest_Update extends Base {
 	 * @return bool|array $response
 	 */
 	private function parse_bitbucket_webhook( $request_body ) {
-		Singleton::get_instance( 'Basic_Auth_Loader', static::$options )->load_authentication_hooks();
+		Singleton::get_instance( 'Basic_Auth_Loader', $this, static::$options )->load_authentication_hooks();
 
 		$request_data = json_decode( $request_body, true );
 
@@ -369,7 +369,7 @@ class Rest_Update extends Base {
 
 		//$response['payload'] = $new;
 
-		Singleton::get_instance( 'Basic_Auth_Loader', static::$options )->remove_authentication_hooks();
+		Singleton::get_instance( 'Basic_Auth_Loader', $this, static::$options )->remove_authentication_hooks();
 
 		return $response;
 	}