+++ src/Singleton.php
@@ -29,26 +29,26 @@ if ( ! class_exists( 'Fragen\\Singleton' ) ) {
 	final class Singleton {
 
 		/**
+		 * Get instance of class.
+		 *
 		 * @param string               $class_name
+		 * @param object               $caller
 		 * @param null|array|\stdClass $options
 		 *
-		 * @return array
+		 * @return array $instance
 		 */
-		public static function get_instance( $class_name, $options = null ) {
+		public static function get_instance( $class_name, $caller = null, $options = null ) {
 			static $instance = null;
 
-			// TODO shorten syntax for PHP 5.4
-			$backtrace = debug_backtrace();
-			$backtrace = $backtrace[1];
-			$class     = isset( $backtrace['class'] ) ? $backtrace['class'] : null;
-			$class     = self::get_class( $class_name, $class );
+			$class = get_class( $caller );
+			$class = self::get_class( $class_name, $class );
 
 			if ( null === $instance || ! isset( $instance[ $class ] ) ) {
 				$instance[ $class ] = new $class( $options );
 			}
 
 			// Add calling object.
-			$instance[ $class ]->caller = isset( $backtrace['object'] ) ? $backtrace['object'] : null;
+			$instance[ $class ]->caller = $caller;
 
 			return $instance[ $class ];
 		}
@@ -59,20 +59,43 @@ if ( ! class_exists( 'Fragen\\Singleton' ) ) {
 		 * @param string $class_name
 		 * @param string $class
 		 *
-		 * @return string $class
+		 * @return string Namespaced class name.
 		 */
 		private static function get_class( $class_name, $class ) {
+			$reflection      = self::get_reflection( $class );
+			$namespace       = $reflection->getNamespaceName();
+			$namespace_parts = explode( '\\', $namespace );
+			$count           = count( $namespace_parts );
+			$classes[ - 1 ]  = null;
+
+			for ( $i = 0; $i < $count; $i ++ ) {
+				$classes[ $i ] = ltrim( $classes[ $i - 1 ] . '\\' . $namespace_parts[ $i ], '\\' );
+			}
+
+			$classes = array_reverse( $classes );
+			foreach ( $classes as $namespace ) {
+				$namespaced_class = $namespace . '\\' . $class_name;
+				if ( class_exists( $namespaced_class ) ) {
+					return $namespaced_class;
+				}
+			}
+		}
+
+		/**
+		 * Get ReflectionClass of passed class name.
+		 *
+		 * @param string $class
+		 *
+		 * @return \ReflectionClass $reflection
+		 */
+		private static function get_reflection( $class ) {
 			try {
-				$reflection      = new \ReflectionClass( $class );
-				$namespace       = $reflection->getNamespaceName();
-				$namespace_parts = explode( '\\', $namespace );
-				$namespace_root  = $namespace_parts[0] . '\\' . $namespace_parts[1];
-				$class           = $namespace_root . '\\' . $class_name;
+				$reflection = new \ReflectionClass( $class );
 			} catch ( \ReflectionException $Exception ) {
 				die( '<table>' . $Exception->xdebug_message . '</table>' );
 			}
 
-			return $class;
+			return $reflection;
 		}
 	}
 }